Move `PFHSignature125` to feature types header
